--- Source/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext"
-  },
-  "files": ["opn.ts"]
-}
